requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
rapidfuzz==3.5.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pandas==2.1.4
requests==2.31.0
psutil==5.9.6
aioredis==2.0.1